add product

+++ app/domain/product/Product.scala
@@ -2,4 +2,4 @@ package domain.product
 
 case class Product(id: Option[Long],
                 name: String,
-                price: Long)
+                price: Float)
